@@ -21,24 +21,25 @@
   "require": {
     "php": ">=8.1",
     "psr/http-client": "^1.0",
-    "psr/log": "^2.0||^3.0",
-    "twig/twig": "^2.0||^3.0"
+    "psr/log": "^2.0 || ^3.0",
+    "twig/twig": "^3.0"
   },
   "require-dev": {
     "dg/bypass-finals": "^1.4",
     "doctrine/dbal": "^4.0",
     "doctrine/migrations": "^3.3",
-    "elasticsearch/elasticsearch": "^6.0||^7.0||^8.0",
-    "mongodb/mongodb": "^1.4",
+    "elasticsearch/elasticsearch": "^8.0",
+    "mongodb/mongodb": "^1.11",
     "monolog/monolog": "^3.0",
     "nyholm/psr7": "^1.5",
     "phpstan/phpstan": "^1.0",
     "phpunit/phpunit": "^10.0",
     "rector/rector": "^1.2",
-    "symfony/browser-kit": "^5.0||^6.0",
-    "symfony/css-selector": "^5.0||^6.0",
-    "symfony/event-dispatcher": "^5.0||^6.0",
-    "symfony/http-client": "^5.0||^6.0"
+    "symfony/browser-kit": "^6.4 || ^7.0",
+    "symfony/css-selector": "^6.4 || ^7.0",
+    "symfony/dom-crawler": "^6.4 || ^7.0",
+    "symfony/event-dispatcher": "^6.4||^7.0",
+    "symfony/http-client": "^6.4||^7.0"
   },
   "suggest": {
     "doctrine/dbal": "for checking DB connections",
